+++ app/src/main/java/com/example/petbutler/ui/Classes/Pessoa/Butler.java
@@ -6,7 +6,7 @@ public class Butler extends Pessoa{
     private String CNH;
     private Conta conta;
 
-    public Butler(String nome, String login, String senha, String CPF, String resumo, float nota, Foto foto, Telefone telefone) {
+    public Butler(String nome, String login, String senha, String CPF, String resumo, double nota, Foto foto, Telefone telefone) {
         super(nome, login, senha, CPF, resumo, nota, foto, telefone);
     }
